Skybox support (#642)
* Added skybox support

Added support for skyboxes under the Resources folder. This will select
a random skybox on initial world creation and persist it to the
world.xml file. I only included 2 skyboxes, but that accounts for 32
files. The skyboxes are from Hedgehog Team's Skyboxes Volume 2 asset.

* Background Camera

Added background camera to the scene as the orthographic projection on
the main camera was messing with the perspective of the skyboxes.

* Merged SCENE.unity

Merged latest SCENE.unity to add background camera.