Makes the Scheduler class visible to Lua scripts
Previously there was no way for Lua scripts to access the Scheduler
directly. They could create or modify ScheduledEvent objects but not
actually (de)schedule them! This was a simple oversight on my part. The
intent for Lua access was there from the beginning.

Fixing this was simply a matter of adding a [MoonSharpUserData]
declaration and registering the globals. Now any Lua function run through
LuaFunctions.Call() or CallWithInstance() automatically gets the Scheduler
class and (most importantly) Scheduler.Current in its globals.

On commit all tests and Stylecop were passing.