Really removes UIController (#1356)
PR #1353 removed the UIController.cs file which contained only an empty
class which was not doing anything in the actual code. This was part of
a reorganisation and simplification of the folder structure for the
controllers. However, not all references to UIController were actually
tidied.

In particular, there was a prefab hanging around and a single reference to
it in WorldController. These have now been cleaned up. There are no
changes to functionality.