Hotfix for GameEvent double up (#1370)
The two file reorganizations in #1366 and #1328 conflicted, resulting in a
double up of GameEvent.cs leading to a compile error. Somehow it didn't
show as a merge conflict so nothing was visible when testing out the PRs.

I have removed the double and merged the Models/Event directory into
Models/Events.