Add a local "Travis" script useful for a pre-commit or pre-push hook.
Scripts/pseudotravis.sh is closely based on the actual Travis build script
Scripts/test.sh, but runs locally. Before using as a git hook script it
must be configured with your local Unity executable path.

To do so change the localUnity env variable at the top of the script.
To use as a pre-commit or pre-push hook simply copy or symlink to the
relevant place in .git/hooks/.