Fixes the HeadlineGenerator so tests run in Monodevelop
Currently the HeadlineGenerator and its test suite depend on UnityEngine
both for its random number generator and for Debug.Log. This means the
tests run fine in the Unity editor but fail in an IDE like Monodevelop
(and I'm guessing probably VS as well).

This issue is easily fixed by using System.Random and disabling logging
during unit tests.