Fix Lua references to room.GetSize to use room.TileCount (#1407)
PR #1072 made the change in Room and all C# code, but neglected to check
for Lua references. These wound up giving error (c.f. #1406).

This just changes to references to the new name.

Also added a null check in LuaFunctions::CallWithInstance because the
return value of the Lua call can be null.

Fixes #1406.