Create Github Issue and Pull Request templates
These templates are automatically filled in the editor textarea when
someone creates an Issue or Pull request on Github.
The text can be changed or deleted at will by the user, it is there merely
to serve as a prompt to include the pertinent information and format it in
a useful way.